pyproject.toml
[project]
name = "technezis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyquery>=2.0.1",
    "pyquery-xpath>=0.6",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
]
